# 50% speed increase for suggest call