I've done some performance tweaking for the suggest call. No changes to api, no changes to results. Algorithm remains the same. This removes some redundancy in processing.

* Now uses a stripped down version of es6 sets (where available) or a polyfill when required. This reduces the amount of suggestions generate by `edits1` by about a half just by removing duplicates
* When checking the suggested results it now uses `self.checkExact` rather than `self.check` as we will only be looking for exact matches here and `self.check` just introduces extra processing.

Here's some stats from some test words...

**takr**
es6: 52ms
polyfil: 54ms
original: 115ms
*55% faster*

**hoouse**
es6: 134ms
polyfil: 133ms
original: 200ms
*33% faster*

**suggestionsr**
es6: 590ms
polyfil: 598ms
original: 1201ms
*51% faster*

**dsdsfdsfdsfdfsdfsdfsdf**
es6: 3105ms
polyfil: 3116ms
original: 5676ms
*46% faster*