@cfinke 2016-09-20
Thanks for the contribution @Thomas101 and @kofifus, and sorry for the delay in merging.