@cfinke 2016-09-20
I had to revert this -- the quality of the `suggest()` results dropped noticeably. @Thomas101 Was there an intended change to functionality, or was it supposed to be performance changes only?