@Thomas101 2016-09-20
The results should have been the same. Do you have any sample suggestion sets? Were you using node or the browser?