@cfinke 2016-09-25
I was running the tests in the browser via the Chrome extension.

An example: `suggest( 'speling' )` returned `["seeking",  "spieling",  "pelting"]`, which does not seem as helpful as the old `[ "spelling", "spieling", "spewing" ]`.  All of the old suggestions were one modification away from the argument to `suggest()`, but two of the new suggestions are two modifications away.