@cfinke 2017-02-13
I realized today why the suggestions changed -- Typo was using the fact that a word added to the `edits1()` return value multiple times in order to weight that word as a more useful spelling suggestion. The WordList class can't take advantage of this because it uses a hash to store those values.